requirements.txt
pytorch==1.7.1
numpy==1.19.2
scipy==1.5.2
scikit-image==0.17.2
libtiff
tqdm